@@ -1,11 +1,16 @@
 part of pdftron;
 
-// Functions define the names of the functions
+/// Defines the names of the functions.
 class Functions {
   static const getPlatformVersion = "getPlatformVersion";
   static const getVersion = "getVersion";
   static const initialize = "initialize";
   static const openDocument = "openDocument";
+  static const getSavedSignatures = "getSavedSignatures";
+  static const getSavedSignatureFolder = "getSavedSignatureFolder";
+
+/// Android only
+  static const getSavedSignatureJpgFolder = "getSavedSignatureJpgFolder";
   static const importAnnotations = "importAnnotations";
   static const exportAnnotations = "exportAnnotations";
   static const flattenAnnotations = "flattenAnnotations";
@@ -64,7 +69,7 @@ class Functions {
   static const exitSearchMode = "exitSearchMode";
 }
 
-// Parameters define the parameters of the functions
+/// Defines the parameters of the functions.
 class Parameters {
   static const licenseKey = "licenseKey";
   static const document = "document";
@@ -96,7 +101,7 @@ class Parameters {
   static const matchWholeWord = "matchWholeWord";
 }
 
-// Parameters define the parameters of the events
+/// Defines the parameters of the events.
 class EventParameters {
   static const action = "action";
   static const annotations = "annotations";
@@ -109,7 +114,7 @@ class EventParameters {
   static const pageNumber = "pageNumber";
 }
 
-// Buttons define the various kinds of buttons for the viewer
+/// Defines the various kinds of buttons for the viewer.
 class Buttons {
   static const viewControlsButton = 'viewControlsButton';
   static const freeHandToolButton = 'freeHandToolButton';
@@ -157,14 +162,14 @@ class Buttons {
   static const redo = 'redo';
   static const showFileAttachmentButton = 'showFileAttachmentButton';
 
-  // Android only
+  /// Android only.
   static const editAnnotationToolbarButton = 'editAnnotationToolButton';
   static const saveReducedCopyButton = 'saveReducedCopyButton';
   static const saveCroppedCopyButton = 'saveCroppedCopyButton';
   static const savePasswordCopyButton = 'savePasswordCopyButton';
 }
 
-// Tools define the various kinds of tools for the viewer
+/// Defines the various kinds of tools for the viewer.
 class Tools {
   static const annotationEdit = 'AnnotationEdit';
   static const textSelect = 'TextSelect';
@@ -212,15 +217,17 @@ class Tools {
   static const formCreateComboBoxField = 'FormCreateComboBoxField';
   static const formCreateListBoxField = 'FormCreateListBoxField';
 
-  // iOS only.
+  /// iOS only.
   static const pencilKitDrawing = 'PencilKitDrawing';
 
-  // Android only.
+  /// Android only.
   static const annotationSmartPen = 'AnnotationSmartPen';
+
+  /// Android only.
   static const annotationLasso = 'AnnotationLasso';
 }
 
-// FitModes define all fit modes in the viewer
+/// Defines all fit modes in the viewer.
 class FitModes {
   static const fitPage = 'FitPage';
   static const fitWidth = 'FitWidth';
@@ -228,7 +235,7 @@ class FitModes {
   static const zoom = 'Zoom';
 }
 
-// LayoutModes define all layout modes in the viewer
+/// Defines all layout modes in the viewer.
 class LayoutModes {
   static const single = 'Single';
   static const continuous = 'Continuous';
@@ -238,19 +245,19 @@ class LayoutModes {
   static const facingCoverContinuous = 'FacingCoverContinuous';
 }
 
-// FieldFlags define the property flags for a form field
+/// Defines the property flags for a form field.
 class FieldFlags {
   static const ReadOnly = 0;
   static const Required = 1;
 }
 
-// ThumbnailFilterModes define the filter modes in thumbnail browser
+/// Defines the filter modes in thumbnail browser.
 class ThumbnailFilterModes {
   static const annotated = "annotated";
   static const bookmarked = "bookmarked";
 }
 
-// AnnotationFlags define the flags for any annotation in the document
+/// Defines the flags for any annotation in the document.
 class AnnotationFlags {
   static const hidden = "hidden";
   static const invisible = "invisible";
@@ -264,24 +271,33 @@ class AnnotationFlags {
   static const toggleNoView = "toggleNoView";
 }
 
-// AnnotationProperties define all possible annotation properties
+/// Defines all possible annotation properties.
 class AnnotationProperties {
-  // not markup exclusive
+  /// Not markup exclusive.
   static const rect = "rect";
+
+  /// Not markup exclusive.
   static const contents = "contents";
-  // markup exclusive
-  static const subject = "subject";
-  static const title = "title";
-  static const contentRect = "contentRect";
+
+  /// Not markup exclusive.
   static const rotation = "rotation";
+
+  /// Markup exclusive.
+  static const subject = "subject";
+
+  /// Markup exclusive.
+  static const title = "title";
+
+  /// Markup exclusive.
+  static const contentRect = "contentRect";
 }
 
-// Behaviors define all user behaviors in the viewer
+/// Defines all user behaviors in the viewer.
 class Behaviors {
   static const linkPress = "linkPress";
 }
 
-// LongPressMenuItems define all menu items that could show up on long press of text or empty area
+/// Defines all menu items that could show up on long press of text or empty area.
 class LongPressMenuItems {
   static const copy = "copy";
   static const search = "search";
@@ -289,7 +305,7 @@ class LongPressMenuItems {
   static const read = "read";
 }
 
-// AnnotationMenuItems define all menu items that could show up on press of annotation
+/// Defines all menu items that could show up on press of annotation.
 class AnnotationMenuItems {
   static const style = "style";
   static const note = "note";
@@ -310,7 +326,7 @@ class AnnotationMenuItems {
   static const ungroup = "ungroup";
 }
 
-// DefaultToolbars define a set of pre-designed toolbars for easier customization
+/// Defines a set of pre-designed toolbars for easier customization.
 class DefaultToolbars {
   static const view = "PDFTron_View";
   static const annotate = "PDFTron_Annotate";
@@ -324,7 +340,7 @@ class DefaultToolbars {
   static const favorite = "PDFTron_Favorite";
 }
 
-// ToolbarIcons define default toolbar icons for use for potential custom toolbars
+/// Defines default toolbar icons for use for potential custom toolbars.
 class ToolbarIcons {
   static const view = "PDFTron_View";
   static const annotate = "PDFTron_Annotate";
@@ -338,13 +354,16 @@ class ToolbarIcons {
   static const favorite = "PDFTron_Favorite";
 }
 
+/// Defines the export format for pdf image
 class ExportFormat {
   static const BMP = "BMP";
   static const JPEG = "JPEG";
   static const PNG = "PNG";
 }
 
-// PTOrientation defines the screen orientations for the viewer. Android only.
+/// Defines the screen orientations for the viewer.
+///
+/// Android only.
 class PTOrientation {
   static const unspecified = -1;
   static const landscape = 0;
@@ -353,24 +372,29 @@ class PTOrientation {
   static const sensorPortrait = 7;
   static const reverseLandscape = 8;
   static const reversePortrait = 9;
-  static const userLandscape =
-      11; // Only changes direction if user has enabled sensor-based rotation.
-  static const userPortrait =
-      12; // Only changes direction if user has enabled sensor-based rotation.
+
+  /// Only changes direction if user has enabled sensor-based rotation.
+  static const userLandscape = 11;
+
+  /// Only changes direction if user has enabled sensor-based rotation.
+  static const userPortrait = 12;
 }
 
-// ViewModePickerItem defines the items in the view mode dialog.
+/// Defines the items in the view mode dialog.
 class ViewModePickerItem {
   static const Crop = "viewModeCrop";
   static const Rotation = "viewModeRotation";
   static const ColorMode = "viewModeColorMode";
+  static const VerticalScrolling = "viewModeVerticalScrolling";
 }
 
-// DefaultEraserType defines the default behaviour of the eraser tool.
+/// Defines the type of eraser used in the viewer.
 class DefaultEraserType {
   static const annotationEraser = "annotationEraser";
-  static const hybrideEraser = "hybrideEraser";
-  static const inkEraser = "inkEraser"; // Android only
+  static const hybrideEraser = "hybridEraser";
+
+  /// Android only.
+  static const inkEraser = "inkEraser";
 }
 
 // ReflowOrientation defines the scrolling direction of the viewer while in reflow mode. Android only.
@@ -386,7 +410,7 @@ class AnnotationManagerEditMode {
 
 class AnnotationManagerUndoMode {
   static const Own = "undoModeOwn";
-  static const All = "unodModeAll";
+  static const All = "undoModeAll";
 }
 
 class ToolbarAlignment {
